fix overflow bugs ex01

+++ 08/08/ex01/span.cpp
@@ -6,7 +6,7 @@
 /*   By: rmander <rmander@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/15 21:39:28 by rmander           #+#    #+#             */
-/*   Updated: 2022/01/16 18:02:03 by rmander          ###   ########.fr       */
+/*   Updated: 2022/01/16 20:07:04 by rmander          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,10 +44,10 @@ long Span::shortestSpan(void) const {
     throw InsufficientDataException();
   Span::const_iter_type ib = _data.begin();
   Span::const_iter_type ie = _data.end();
-  long min = std::abs(*(ib + 1) - *ib);
+  long min = std::abs(static_cast<long>(*(ib + 1)) - static_cast<long>(*ib));
   long curr = min;
   for (Span::const_iter_type it = ib + 2; it != ie; ++it) {
-    curr = std::abs(*it - *(it - 1));
+    curr = std::abs(static_cast<long>(*it) - static_cast<long>(*(it - 1)));
     if (curr <= min) {
       min = curr;
     }
